Initial refactor, more work needed
